fitur 4 done

+++ src/main/java/com/apap/tugas1/service/InstansiService.java
@@ -3,9 +3,10 @@ package com.apap.tugas1.service;
 import com.apap.tugas1.model.InstansiModel;
 
 import java.util.List;
+import java.math.BigInteger;
 
 public interface InstansiService {
-	InstansiModel getDetailById(Long id);
+	InstansiModel getDetailById(BigInteger id);
 	List<InstansiModel> getAllInstansi();
 }
