+++ src/main/java/com/apap/tugas1/service/ProvinsiService.java
@@ -3,9 +3,10 @@ package com.apap.tugas1.service;
 import com.apap.tugas1.model.ProvinsiModel;
 
 import java.util.List;
+import java.math.BigInteger;
 
 public interface ProvinsiService {
-	ProvinsiModel getDetailById(Long id);
+	ProvinsiModel getDetailById(BigInteger id);
 	List<ProvinsiModel> getAllProvinsi();
 }
